# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "What to watch API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User registration",
                "description": "Process registration",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of films",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "description": "Process login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of films",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "description": "Process logout",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Successful logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Токены удалены"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginUserRequest": {
                "properties": {
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "ivan@somemail.ru",
                        "nullable": false
                    },
                    "password": {
                        "description": "Password",
                        "type": "string",
                        "example": "ud5R#gh78Qi!6",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
            "RegisterUserRequest": {
                "properties": {
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "Ivan",
                        "nullable": false
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "ivan@somemail.ru",
                        "nullable": false
                    },
                    "password": {
                        "description": "Password",
                        "type": "string",
                        "example": "ud5R#gh78Qi!6",
                        "nullable": false
                    },
                    "password_confirmation": {
                        "description": "Password confirmation",
                        "type": "string",
                        "example": "ud5R#gh78Qi!6",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
            "TokenResource": {
                "title": "TokenResource",
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "0faa527e-deb5-4e92-84f5-b08355d88350|c5O42jaXgOtaSin7KvU9EhF73jXlcFuwFpwOfZE2b4c34717"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Bearer token for authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}